rename Sig => Sg; fix up pretty printer (details below)
Sig => Sg just for consistency with Pi

the changes in the pretty printer are both cosmetic and substantive. The
cosmetic part is to use that %a thing to print each constructor in one
go. The substantive part is that the hov-boxes were over-lapping with
parentheses in a non-projective way, so I fixed that.