improve sharing in implementation of hcom_path (details below)
We can prevent pieces of syntax like (ap fam [i]) and (boundary i) from
being evaluated again and again by adding an explicit let binding in the
syntax/term we are building up. Then, it is evaluated once, added to the
environment, and shared in the remaining syntax as a variable.