Random cleanup (#73)
* starting to deal with source locations

* cleanup

* pushing around the span

* source positions printing, boom

* more

* cleanup

* Beginning of Thought beam.

* Thought transferred from #70.

I did not do n-ary join/meet because I'm not sure about the
associativity of the location information. Maybe it is better
to have built-in join/meet and kill top/bot?

* rename "quasiquote" stuff to "splice"

* move symbol stuff to basis

* first cut at splitting up Nbe

* renaming Nbe to Conversion, etc.

* clean up interface

* error hygiene

* cleaning up interfaces

* more cleanup

* tweak

* cleaning up span stuff

* a very crappy "user-friendly" pretty printer (does not resolve #66)

* much more powerful cofibration normalization

* N-ary join/meet.

I tried to keep the logic of the current code as much as possible,
which means it deserves further clean-ups. Also, a module Util is
actively used in various modules under the name MU. The organization
also needs further refactoring to remove the ugly "MU." I think we can
introduce something like this:

module M = Monad.WithPowerPack (M)

where the functor WithPowerPack opens M again.

* Even better grammar rule names.

* nat-path

Co-authored-by: favonia <favonia@gmail.com>