first cut at adding import to the grammar. it compiles but does not work, so this is partial work, i'm just cacheing it so i don't lose it.
i am trying to do something that seems simple but may be a little fancy,
which is that i want to include statements like

import a
import b.c.d

i believe the right way to do this is to define a syntactic class in the
lexer for paths and then use the built-in non-empty separated list function
so that these parse directly as

Import(["a"])
Import(["b","c","d"])

respectively, where the lexer enforces that each element of the . delimited
qualified name is from [a-zA-Z0-9]. i am chasing my own tail around trying
to write that.

the next question that i'm not sure about is that the identifiers that
appear in atoms throughout will (likely) need to refer to qualified names,
e.g. `data.nat.Nat` for the type of natural numbers defined in the
library. that's going to be a change to the syntax of atoms, but won't make
atoms the same as paths. i'll need to figure that out too.